installed material-ui and fixed react folders
