+++ client/skill_lab/src/Components/App/App.js
@@ -1,0 +1,16 @@
+import "./App.css";
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        Learn React
+      </header>
+    </div>
+  );
+}
+
+export default App;
